Split up AltUtils
