+++ Wurst Client/src/tk/wurst_client/files/FileManager.java
@@ -27,7 +27,6 @@ import org.lwjgl.input.Keyboard;
 
 import tk.wurst_client.Client;
 import tk.wurst_client.alts.Alt;
-import tk.wurst_client.encryption_api.Encryption;
 import tk.wurst_client.gui.alts.GuiAltList;
 import tk.wurst_client.module.Category;
 import tk.wurst_client.module.Module;
